+++ app/Filament/Clusters/Courier.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Filament\Clusters;
+
+use BackedEnum;
+use Filament\Clusters\Cluster;
+use Filament\Pages\Enums\SubNavigationPosition;
+use Filament\Support\Icons\Heroicon;
+
+class Courier extends Cluster
+{
+    protected static string|BackedEnum|null $navigationIcon = Heroicon::OutlinedTruck;
+
+    protected static ?string $navigationLabel = 'Courier';
+
+    protected static ?int $navigationSort = 1;
+
+    protected static ?SubNavigationPosition $subNavigationPosition = SubNavigationPosition::Top;
+}